pyproject.toml
[tool.poetry]
name = "cribot"
version = "0.1.0"
description = ""
authors = ["arganavtus <arganavtus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
pydantic = "^2.9.2"
aiogram = "^3.14.0"
betterlogging = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
